refactor: decompose coverageTransformer into focused modules
- Move TransformOptions, CoverageInput, LineTotals, ProcessResult to utils/types.ts
- Extract tryReadJson to utils/readJson.ts
- Extract processDeployCoverage, processTestCoverage, createHandlers,
  countLines to transformers/coverageProcessor.ts
- coverageTransformer.ts is now a thin orchestrator
